docs: run cem on each build
